// src/store/collectionFeed/collectionFeed.enum.ts
export enum ActionType {
  FETCH_COLLECTION_FEED_SUCCESS = "FETCH_COLLECTION_FEED_SUCCESS",
  FETCH_COLLECTION_FEED_PENDING = "FETCH_COLLECTION_FEED_PENDING",
  FETCH_COLLECTION_FEED_ERROR = "FETCH_COLLECTION_FEED_ERROR",
  FETCH_COLLECTION_DETAIL_SUCCESS = "FETCH_COLLECTION_DETAIL_SUCCESS",
  FETCH_COLLECTION_DETAIL_PENDING = "FETCH_COLLECTION_DETAIL_PENDING",
  RESET_COLLECTION_STATE = "RESET_COLLECTION_STATE",
  INCREMENT_COLLECTION_PAGE = "INCREMENT_COLLECTION_PAGE",
}